Removed unnecesary debug logging.
While testing i had some parts of the shell script log certain operations for debugging purposes. The test version had an `mkdir` command in order to create a directory to save those logs. Since the release version did not require this kind of logging to take place i had removed the command responsible for creating the log directory, however i forgot to remove the commands that appended information to a file that now(according to the script) was supposed to go into a directory that did not exist. Which resulted in an error. That's since been fixed.